+++ .bob/notes/pending-notes.txt
@@ -1,3 +1,2 @@
-{"id":"7a7d9a1b-85de-4483-8732-008a4b22a572","ts":"2026-04-17T16:59:31.240Z","path":"/Users/cedricklunven/dev/astra-db-java/integrations/data-api-spring-boot-3x-starter/src/main/java/com/datastax/astra/spring/DataApiCollectionCrudRepository.java","version":"1.0.0","taskID":"24aa1015-e426-48cc-9669-8e6b15e362b4"}
-{"id":"dc628fad-b328-4023-8e61-4d8126e03128","ts":"2026-04-17T18:22:56.718Z","path":"/Users/cedricklunven/dev/astra-db-java/astra-db-java/src/main/java/com/datastax/astra/internal/serdes/collections/DataAPIAnnotationIntrospector.java","version":"1.0.0","taskID":"24aa1015-e426-48cc-9669-8e6b15e362b4"}
-{"id":"ea271fa1-98f8-4eec-83af-5182764d0561","ts":"2026-04-17T18:24:29.870Z","path":"/Users/cedricklunven/dev/astra-db-java/astra-db-java/src/main/java/com/datastax/astra/internal/serdes/collections/DocumentSerializer.java","version":"1.0.0","taskID":"24aa1015-e426-48cc-9669-8e6b15e362b4"}
+{"id":"a444944b-d1d4-4a03-814f-0dcd321fd968","ts":"2026-04-17T20:37:20.183Z","path":"/Users/cedricklunven/dev/astra-db-java/samples/sample-spring-boot3x/src/main/java/com/ibm/astra/demo/controller/HomeController.java","version":"1.0.0","taskID":"24aa1015-e426-48cc-9669-8e6b15e362b4"}
+{"id":"46e8eaeb-0c6e-416d-885f-d17c36bfa652","ts":"2026-04-17T21:26:50.143Z","path":"/Users/cedricklunven/dev/astra-db-java/samples/sample-spring-boot3x/src/main/java/com/ibm/astra/demo/controller/BookController.java","version":"1.0.0","taskID":"24aa1015-e426-48cc-9669-8e6b15e362b4"}
